Create ZillowAPIDataExtractionUpdated.ipynb
Updated the notebook. It now makes two new datasets. Dataset 1 consists of all the information we need per "12 key" property building, basically it contains all the individual unit information for each unit in its according apartment building. Rather than having duplicate information for each unit, Dataset 2- consists of all specific building information and an overall of its units, I set it up so that there are many ways to link the tables in a schema, this can be done by using- address, geolocation, lotId, or buildingId if needed. Note - zpid is the individual unit identifier, not the apartment building identifier, that is the lotid