Beginning work with Breadcrumbs
Got started with breadcrumbs. Had to do some adjustments to the routes. I'm not too happy with the solution at the moment, but I think with a review + some reworks it lays some good groundwork.

Had to do some research on react router to figure out how the hooks work. Main thing to note are the useNavigate and useLocation hooks.