Created New Logo
I created a new simplified logo and centered it on the navbar, which is also responsive. I'm not too happy with the logo, it's not exactly pretty, but it works.

I also refactored the FilterFields component out of the search and filter. so that it can be reused on another page.