Create cleaning_2.ipynb
My progress on cleaning and preparing the data for a table schema. Lot's of planning and getting the balance between normalization and simplicity had to be done. I tried my best to reduce redundancy between datasets as much as possible, while creating a schema that will require the least amount of joins and the most amount of straight forwardness as possible. Time spent working today - 8 hours 25 minutes