Modifying State Changes
In order to maintain correct and consistent state for the active tags, I had to do some reworking with how they are displayed. This code definitely will require some refactoring, as it introduces a lot of dependency between parent + child components for the filters, however, this can wait until we get features covered.

Main thing to know is that the filterFields component relies on a parent component providing it with a form context provider.

For the moment, active tags cannot be clicked to deactivate them, however, this functionality could be worked in later when there is less pressure.