Setup a Test with API Gateway and Amplify
I was able to set up a test lambda function and a test API gateway. By doing this I was able to understand the process for connecting to the backend. Unfortunately, there is a blocker to continuiing until CORS is enabled on the backend, but once that issue is sorted, it should prove more straightforward to connect to the backend.