Auth Token Refresh Issue
I came across an issue with the authentication tokens due to them expiring after only 60 mins.  The refresh token was not being used automatically. To fix this, I created a getAuthToken function and passed it as part of the UserContextProvider, which returns the authentication token using the Auth library from amplify.

I swapped updated some mutations/queries to use this new way of grabbing the auth token.