Create UserCollaborativeFiltering.ipynb
Wrote the code for the user collaborative filtering recommendation system, it works but needs to be evaluated. Since, it was made using mock data we need to test it with randomized data instead. I populated the dataset using a randomized skewed distribution but this obviously still isn't real. Next I will upload the mock data to the database until we can populate it with actual data