Create cosineSimilarityHybrid4.py
Instead of using one hot encodings and the property detail information. I realised that all of this information was noisy. The returned recommendations were not similar in the key information such as price, location, and bedroom count, which should be the most similar information in recommended properties as its what the users will want to see. Click data is incorporated too, which will allow for the more niche similarities to be recommended if the user clearly desires them