Create cosineSimilarityHybrid7.py
This hybrid model combines property content with user click data, however the results are returning very high cosine angles, this means that the the formula isn't finding similar vectors. When I remove the property content information, the cosine off the angle between the vectors and the input vector goes right down, indicating that the property data is creating noise, and that we have a curse of dimensionality problem. For this reason, we will use the non hybrid during user testing, and see how it compares to the knn model that does include property content data. The results of the cosine similarity hybrid are so poor that this model can be disregarded